Refactoring the z-indexes scss to match naming conventions with the others under Fractal.
